% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Plot the location of accidents on US maps}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{Integers or numerical values for a US State numbers included in the analysis}

\item{year}{An integer or numerical value for the year of the analysis}
}
\value{
a US map woith the location of accidents for the year specified in the parameters
Returns \code{invalid STATE number} if the state number(s) can not be found in the
file or directory
Returns \code{no accidents to plot} if there are no accidents in the chosen file
for the analysis
}
\description{
This function load accident data from
the US National Highway Traffic Safety Administration's Fatality Analysis Reporting System,
into R for a specified year and state numbers,
validate the parameters and plot the location of accidents on a US map.
}
\examples{
\dontrun{
fars_map_state(c(1:56),2014)
}
}
